add pager
add pager

+++ idl/common.thrift
@@ -9,12 +9,12 @@ struct BaseResponse {
 struct NilResponse {}
 
 struct Pager {
-    1: i32 paged,
-    2: i32 total,
-    3: i32 page_count,
-    4: i32 page_size,
-    5: i32 prev_page,
-    6: i32 last_page,
+    1: i64 paged,
+    2: i64 total,
+    3: i64 page_count,
+    4: i64 page_size,
+    5: i64 prev_page,
+    6: i64 last_page,
 }
 
 enum BoolStatus {
